src/diabetes/requirements.txt
azureml-mlflow==1.47.0
mlflow-skinny==1.30.0
numpy==1.23.4
pandas==1.5.1
matplotlib==3.6.2
sklearn==0.0
scikit-learn==1.1.1
